Creates a utils folder and moves exceptions folder into model
