+++ core/src/main/se/tevej/game/model/ModelManager.java
@@ -5,8 +5,8 @@ import com.badlogic.ashley.core.Entity;
 import com.badlogic.ashley.core.EntityListener;
 import com.badlogic.ashley.signals.Signal;
 
-import main.se.tevej.game.Manager;
-import main.se.tevej.game.Options;
+import main.se.tevej.game.utils.Manager;
+import main.se.tevej.game.utils.Options;
 import main.se.tevej.game.model.ashley.SignalListener;
 import main.se.tevej.game.model.components.InventoryComponent;
 import main.se.tevej.game.model.entities.WorldEntity;
