+++ core/src/main/se/tevej/game/model/components/InventoryComponent.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import com.badlogic.ashley.core.Component;
 
-import main.se.tevej.game.exceptions.NotEnoughResourcesException;
+import main.se.tevej.game.model.exceptions.NotEnoughResourcesException;
 import main.se.tevej.game.model.utils.Resource;
 import main.se.tevej.game.model.utils.ResourceType;
