+++ core/src/main/se/tevej/game/model/exceptions/MissmatchedResourceException.java
@@ -1,4 +1,4 @@
-package main.se.tevej.game.exceptions;
+package main.se.tevej.game.model.exceptions;
 
 public class MissmatchedResourceException extends Exception {
     public MissmatchedResourceException() {
