+++ core/src/main/se/tevej/game/model/exceptions/UnknownResourceException.java
@@ -1,4 +1,4 @@
-package main.se.tevej.game.exceptions;
+package main.se.tevej.game.model.exceptions;
 
 import main.se.tevej.game.model.utils.ResourceType;
