+++ core/src/main/se/tevej/game/model/systems/BuildBuildingSystem.java
@@ -6,7 +6,7 @@ import com.badlogic.ashley.core.EntitySystem;
 import com.badlogic.ashley.signals.Listener;
 import com.badlogic.ashley.signals.Signal;
 
-import main.se.tevej.game.exceptions.NoSuchBuildingException;
+import main.se.tevej.game.model.exceptions.NoSuchBuildingException;
 import main.se.tevej.game.model.ashley.SignalComponent;
 import main.se.tevej.game.model.ashley.SignalListener;
 import main.se.tevej.game.model.components.PositionComponent;
